+++ app/charts/line/chart-lines.tsx
@@ -33,7 +33,7 @@ import {
 import { isNumber } from "@/configurator/components/ui-helpers";
 import { Observation } from "@/domain/data";
 import {
-  DimensionMetaDataFragment,
+  DimensionMetadataFragment,
   useDataCubeObservationsQuery,
 } from "@/graphql/query-hooks";
 import { useLocale } from "@/locales/use-locale";
@@ -103,8 +103,8 @@ export const ChartLines = memo(function ChartLines({
   interactiveFiltersConfig,
 }: {
   observations: Observation[];
-  dimensions: DimensionMetaDataFragment[];
-  measures: DimensionMetaDataFragment[];
+  dimensions: DimensionMetadataFragment[];
+  measures: DimensionMetadataFragment[];
   fields: LineFields;
   interactiveFiltersConfig: InteractiveFiltersConfig;
 }) {