+++ app/charts/shared/chart-helpers.tsx
@@ -21,7 +21,7 @@ import {
 import { FIELD_VALUE_NONE } from "@/configurator/constants";
 import { Observation } from "@/domain/data";
 import { truthy } from "@/domain/types";
-import { DimensionMetaDataFragment } from "@/graphql/query-hooks";
+import { DimensionMetadataFragment } from "@/graphql/query-hooks";
 
 export type QueryFilters = Filters | FilterValueSingle;
 
@@ -286,7 +286,7 @@ const SlugRe = /\W+/g;
 export const getSlugifiedIri = (iri: string) => iri.replace(SlugRe, "_");
 
 export const getLabelWithUnit = (
-  dimension: DimensionMetaDataFragment
+  dimension: DimensionMetadataFragment
 ): string => {
   return dimension.unit
     ? `${dimension.label} (${dimension.unit})`