+++ app/configurator/components/chart-controls/color-palette.tsx
@@ -24,14 +24,14 @@ import {
   getPalette,
   mapValueIrisToColor,
 } from "@/configurator/components/ui-helpers";
-import { DimensionMetaDataFragment } from "@/graphql/query-hooks";
+import { DimensionMetadataFragment } from "@/graphql/query-hooks";
 import useEvent from "@/utils/use-event";
 
 type Props = {
   field: string;
   disabled?: boolean;
   colorConfigPath?: string;
-  component: DimensionMetaDataFragment | undefined;
+  component: DimensionMetadataFragment | undefined;
 };
 
 const useStyles = makeStyles({
@@ -180,7 +180,7 @@ const ColorPaletteReset = ({
 }: {
   field: string;
   colorConfigPath?: string;
-  component: DimensionMetaDataFragment;
+  component: DimensionMetadataFragment;
   state: ConfiguratorStateConfiguringChart;
 }) => {
   const [, dispatch] = useConfiguratorState();